Make saving to SRAM work
SRAM needs to be enabled before it's usable.
Best practice is to then disable it again rather than leaving it on,
as otherwise you can get random writes during poweroff (or so it's claimed).

So we keep ReadTail in RAM and only copy it to SRAM when we change it.